+++ src/main/scala/com/atomist/rug/functions/travis/RestartBuildFunction.scala
@@ -17,29 +17,20 @@ class RestartBuildFunction
   /**
     * Restart a presumably failed Travis CI build.
     *
-    * @param visibility   whether repo/build is "public" or "private", used to determine which Travis API to hit
+    * @param owner        GitHub owner, i.e., user or organization, of the repo to enable
+    * @param repo         name of the repo to enable
     * @param buildId      ID of build to restart
     * @param githubToken  GitHub token with proper scopes for Travis CI
     * @return
     */
   @RugFunction(name = "restart-travis-build", description = "Restarts a travis build",
     tags = Array(new Tag(name = "travis"), new Tag(name = "ci")))
-  def invoke(@Parameter(name = "visibility") visibility: String,
+  def invoke(@Parameter(name = "owner") owner: String,
+             @Parameter(name = "repo") repo: String,
              @Parameter(name = "buildId") buildId: String,
              @Secret(name = "githubToken", path = TravisFunction.githubTokenPath) githubToken: String): FunctionResponse = {
-
-      val api: TravisAPIEndpoint = TravisAPIEndpoint.stringToTravisEndpoint(visibility)
-      val travisToken = travisEndpoints.postAuthGitHub(api, GitHubToken(githubToken))
-      val headers: HttpHeaders = TravisEndpoints.authHeaders(travisToken)
-      try {
-        travisEndpoints.postRestartBuild(api, headers, buildId)
-        FunctionResponse(Status.Success, Option(s"Successfully restarted build `$buildId` on Travis CI"), None, None)
-      }
-      catch {
-        case e: Exception =>
-          logger.error(s"$visibility build $buildId restart failed: ${e.getMessage}", e)
-          FunctionResponse(Status.Failure, Some(s"Failed to restart build `$buildId` on Travis CI"), None, StringBodyOption(e.getMessage))
-      }
+    val repoSlug = RepoSlug(owner, repo)
+    RestartBuild(travisEndpoints, gitHubRepo).tryRestart(repoSlug, buildId, GitHubToken(githubToken))
   }
 }